Refactor: move function back to actions
